# openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeIn": {
        "properties": {
          "facts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Fact"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facts"
          },
          "lang_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang Hint"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "reply": {
            "maxLength": 4000,
            "title": "Reply",
            "type": "string"
          }
        },
        "required": [
          "reply"
        ],
        "title": "AnalyzeIn",
        "type": "object"
      },
...
      "CheckIn": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "title": "Facts",
            "type": "array"
          },
          "lang_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang Hint"
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "reply": {
            "maxLength": 4000,
            "title": "Reply",
            "type": "string"
          }
        },
        "required": [
          "facts",
          "reply"
        ],
        "title": "CheckIn",
        "type": "object"
      },
...
      "Fact": {
        "properties": {
          "critical": {
            "default": true,
            "title": "Critical",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/FactType"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "id",
          "type",
          "value",
          "label"
        ],
        "title": "Fact",
        "type": "object"
      },
      "FactType": {
        "enum": [
          "dose",
          "frequency",
          "timing",
          "duration",
          "date",
          "amount",
          "condition"
        ],
        "title": "FactType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LlmToggle": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "LlmToggle",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Decode: plain English, where / when / what / how much, from accented or mixed speech and messages (text in, text out). Check mode: teach-back for mixed-language messages. Deterministic engine; LLM, speech-to-text and translation are optional helpers. Decode contract: docs/API.md.",
    "title": "Steve API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/analyze": {
      "post": {
        "description": "Stage-by-stage view (tokens, lexicon matches, slots, results) for /how-it-works.",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "engine"
        ]
      }
    },
    "/check": {
      "post": {
        "description": "Stateless: {facts, reply} -> per-fact results. Used by the landing playground and the eval.",
        "operationId": "check_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check",
        "tags": [
          "engine"
        ]
      }
    },
...
    "/demo/seed": {
      "post": {
        "description": "(Re)load data/scenarios.json as confirmed demo messages OWNED BY THE CALLER (ids carry a per-sender suffix).\nThe first one gets a pre-baked 'subtle mistake' reply.",
        "operationId": "demo_seed_demo_seed_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-sender-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sender-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Seed",
        "tags": [
          "demo"
        ]
      }
    },
...
    "/settings/llm": {
      "post": {
        "description": "Runtime toggle for the wrapper test: with it OFF, everything still works. GLOBAL, so admin-only (X-Admin-Key).",
        "operationId": "set_llm_settings_llm_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LlmToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Llm",
        "tags": [
          "settings"
        ]
      }
    }
  }
}